% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemma3.R
\name{gemma3}
\alias{gemma3}
\alias{gemma3_from_config}
\alias{gemma3_from_pretrained}
\title{Gemma3}
\usage{
gemma3(
  vocab_size = 262144,
  n_embd = 640,
  n_head = 4,
  d_head = 256,
  n_inter = 2048,
  max_pos = 32768,
  sliding_window = 512,
  rope_base = c(local = 10000, global = 1e+06),
  layer_types = rep(c(rep("local", 5), "global"), times = 3),
  rmsnorm_eps = 1e-06
)

gemma3_from_config(identifier, revision = "main")

gemma3_from_pretrained(identifier = "google/gemma-3-270m", revision = "main")
}
\description{
Gemma3
}
\section{Functions}{
\itemize{
\item \code{gemma3_from_config()}: Initializes a gemma3 model using a configuration defined in HF Hub

\item \code{gemma3_from_pretrained()}: Initializes the gemma3 model and load pre-trained weights from HF hub.

}}
