% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gptbigcode.R
\name{gptbigcode}
\alias{gptbigcode}
\alias{gptbigcode_from_config}
\alias{gptbigcode_from_pretrained}
\title{GPT BigCode}
\usage{
gptbigcode(
  vocab_size = 50257,
  n_embd = 768,
  n_head = 12,
  n_layer = 12,
  max_pos = 1024,
  pdrop = 0.1
)

gptbigcode_from_config(identifier, revision = "main")

gptbigcode_from_pretrained(identifier, revision = "main")
}
\arguments{
\item{vocab_size}{An integer indicating the size of the vocabulary or the number
of unique tokens in the input data.}

\item{n_embd}{An integer specifying the dimensionality of the embedding vectors.}

\item{n_head}{An integer representing the number of attention heads in the
multi-head attention mechanism.}

\item{n_layer}{An integer indicating the number of layers in the deep learning model.}

\item{max_pos}{An integer specifying the maximum position encoding value or
the maximum sequence length.}

\item{pdrop}{Dropout probability the attention, residual and embeddings dropout.}

\item{identifier}{A string representing the identifier or name of the pre-trained
model in the Hugging Face model hub.}

\item{revision}{A string specifying the revision or version of the pre-trained
model in the Hugging Face model hub.}
}
\value{
An initialized \code{\link[torch:nn_module]{torch::nn_module()}}.
}
\description{
Initializes a BigCode model
}
\section{Functions}{
\itemize{
\item \code{gptbigcode_from_config()}: Initializes a \code{gptbigcode} from a config file from HF hub

\item \code{gptbigcode_from_pretrained()}: Initializes a \code{gptbigcode} and loads pre-trained weights from HF Hub

}}
