% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{hf_state_dict}
\alias{hf_state_dict}
\title{Loads, and possibly downloads HF Hub models}
\usage{
hf_state_dict(identifier, revision = "main")
}
\arguments{
\item{identifier}{Repository id from the model}

\item{revision}{Revision to download from (eg tags, branches or commit hashes)}
}
\description{
Loads, and possibly downloads HF Hub models
}
